+++ bookmark-in-project.el
@@ -82,7 +82,7 @@ A nil return value will fall back to the `default-directory'."
 
 
 ;; ---------------------------------------------------------------------------
-;; Generic Macros
+;; Generic Macros / Functions
 
 (defun bookmark-in-project--canonicalize-path (path)
   "Return the canonical PATH.
@@ -101,6 +101,34 @@ This is done without adjusting trailing slashes or following links."
   (let ((file-name-handler-alist nil))
     ;; Expand user `~' and default directory.
     (expand-file-name path)))
+
+(defun bookmark-in-project--nth-elt (element xs &rest args)
+  "Return zero-indexed position of ELEMENT in list XS, or nil if absent.
+Optional ARGS, may contain :test keyword argument,
+it's value is used instead of `equal' for comparison,
+where the first argument is always ELEMENT, the second is a member of XS."
+  (let
+    (
+      (index 0)
+      (test-fn (or (plist-get args :test) #'equal)))
+    (while (and xs (not (funcall test-fn element (car xs))))
+      (setq index (1+ index))
+      (setq xs (cdr xs)))
+    (and xs index)))
+
+(defun bookmark-in-project--nth-vec (element xs &rest args)
+  "Return zero-indexed position of ELEMENT in vector XS, or nil if absent.
+Optional ARGS, may contain :test keyword argument,
+it's value is used instead of `equal' for comparison,
+where the first argument is always ELEMENT, the second is a member of XS."
+  (let
+    (
+      (xs-length (length xs))
+      (index 0)
+      (test-fn (or (plist-get args :test) #'equal)))
+    (while (and (< index xs-length) (not (funcall test-fn element (aref xs index))))
+      (setq index (1+ index)))
+    (and (< index xs-length) index)))
 
 
 ;; ---------------------------------------------------------------------------
@@ -454,6 +482,18 @@ Argument DIRECTION represents the stepping direction (in -1 1)."
           (-1 (min (1- (point)) (line-beginning-position)))
           (_ (error "Invalid direction")))))))
 
+(defun bookmark-in-project--item-handle-or-nil (item)
+  "Jump to ITEM, returning non-nil on success."
+  (let ((bookmark-name-or-record (car item)))
+    (condition-case _err
+      (progn
+        (funcall
+          (or (bookmark-get-handler bookmark-name-or-record) 'bookmark-default-handler)
+          (bookmark-get-bookmark bookmark-name-or-record))
+        t)
+      (error nil))))
+
+
 (defun bookmark-in-project--item-get-filename (item)
   "Return the filename from a bookmark (ITEM)."
   (let
@@ -484,6 +524,22 @@ otherwise it will switch the buffer."
     (bookmark-handle-bookmark (car item))
     (point)))
 
+;; Extracted from: `bookmark-handle-bookmark',
+;; Doesn't attempt to handle errors, just skip them.
+(defun bookmark-in-project--item-get-position-or-nil (item)
+  "Return the position of bookmark ITEM.
+Note that this must only run on the for bookmarks in the current buffer,
+otherwise it will switch the buffer."
+  ;; Note that edits to the document mean: (alist-get 'position item)
+  ;; May not reflect the location after the context has been used to resolve the actual point.
+  ;; For this reason, the actual jump call is needed.
+  (save-excursion
+    (cond
+      ((bookmark-in-project--item-handle-or-nil item)
+        (point))
+      (t
+        nil))))
+
 (defun bookmark-in-project--compare (a b)
   "Return t when A is less than B."
   (let
@@ -502,24 +558,223 @@ otherwise it will switch the buffer."
       (t
         nil))))
 
-(defun bookmark-in-project--step-index (bm-list direction)
-  "Return the index in the BM-LIST, stepping in DIRECTION direction."
+
+;; ---------------------------------------------------------------------------
+;; Bookmark Pretty Item API
+;;
+;; Use to cache expensive positional lookups for bookmarks.
+
+(defun bookmark-in-project--pretty-item-from-bookmark (item)
+  "Return a pretty bookmark from ITEM: (item . (position . invalid)).
+When the `position' is nil, it's not yet initialized.
+When it's an integer it is, and the value of `invalid' will set."
+  (cons item (cons nil nil)))
+
+(defun bookmark-in-project--pretty-item-ensure-position (pi)
+  "Initialize the position PI as needed."
+  (let ((pi-pos-cell (cdr pi)))
+    ;; Not yet initialized.
+    (when (null (car pi-pos-cell))
+      (let ((item (car pi)))
+        (let ((pos (bookmark-in-project--item-get-position-or-nil item)))
+          (cond
+            (pos
+              ;; Set the position and mark as valid.
+              (setcar pi-pos-cell pos))
+            (t
+              ;; Set the position to the
+              (setcar pi-pos-cell (alist-get 'position item 1))
+              (setcdr pi-pos-cell t))))))))
+
+(defun bookmark-in-project--pretty-item-ensure-position-for-list (pi-list)
+  "Ensure all items in PI-LIST have their positions initialized."
+  (while pi-list
+    (let ((pi (pop pi-list)))
+      (bookmark-in-project--pretty-item-ensure-position pi))))
+
+(defun bookmark-in-project--pretty-items-from-list (bm-list)
+  "Return a list of pretty items from BM-LIST."
+  (mapcar #'bookmark-in-project--pretty-item-from-bookmark bm-list))
+
+(defun bookmark-in-project--pretty-items-by-file-vector (pi-list &optional extra-files)
+  "Return a sorted vector of `(file-path . pi-list)' pairs from PI-LIST.
+Optionally include EXTRA-FILES (dummy files useful for ordering)."
   (let
     (
-      (len (length bm-list))
-      (item-placeholder (bookmark-in-project--placeholder-item direction))
-      (result 0))
-    (while bm-list
-      (let ((item (car bm-list)))
+      (filepath-pi-list-pairs nil)
+      (files-map (make-hash-table :test #'equal)))
+    (while extra-files
+      (puthash (pop extra-files) nil files-map))
+    (while pi-list
+      (let*
+        (
+          (pi (pop pi-list))
+          (filepath-iter (bookmark-in-project--item-get-filename (car pi)))
+          (pi-list-local (gethash filepath-iter files-map)))
+
+        ;; Maintain a list of items.
+        (puthash filepath-iter (cons pi pi-list-local) files-map)))
+
+    (maphash (lambda (key value) (push (cons key value) filepath-pi-list-pairs)) files-map)
+    (vconcat (sort filepath-pi-list-pairs (lambda (a b) (string-lessp (car a) (car b)))))))
+
+
+;; ---------------------------------------------------------------------------
+;; Bookmark Relative Navigation Next/Previous
+
+(defun bookmark-in-project--pi-list-sorted-by-pos (pi-list)
+  "Return PI-LIST sorted by position."
+  (bookmark-in-project--pretty-item-ensure-position-for-list pi-list)
+  ;; Sort by fallback position.
+  (sort pi-list (lambda (a b) (< (cadr a) (cadr b)))))
+
+(defun bookmark-in-project--step-in-buffer (pi-list direction filepath-current pos-current)
+  "Return the next/previous bookmark based on DIRECTION in PI-LIST.
+Arguments FILEPATH-CURRENT & POS-CURRENT are used as a reference.
+When no bookmark is found in the buffer, return nil."
+  ;; Take care, `filepath-current' and `pos-current' may not be the current buffer
+  ;; so avoid (point-min) or anything that relies on the current buffers values.
+  (let
+    (
+      (pi-list-local (list))
+      (pi-best nil)
+      (pos-best
         (cond
-          ((bookmark-in-project--compare item item-placeholder)
-            (setq bm-list (cdr bm-list))
-            (setq result (1+ result)))
+          ((< direction 0)
+            0)
           (t
-            (setq bm-list nil)))))
-    (when (< direction 0)
-      (setq result (1- result)))
-    (mod result len)))
+            most-positive-fixnum))))
+    (while pi-list
+      (let ((pi (pop pi-list)))
+        (when (string-equal filepath-current (bookmark-in-project--item-get-filename (car pi)))
+          (push pi pi-list-local))))
+
+    (when pi-list-local
+      (setq pi-list-local (bookmark-in-project--pi-list-sorted-by-pos pi-list-local))
+      (while pi-list-local
+        (let*
+          (
+            (pi (pop pi-list-local))
+            (pos (cadr pi)))
+          (when
+            (cond
+              ((< direction 0)
+                (and (< pos-best pos) (> pos-current pos)))
+              (t
+                (and (> pos-best pos) (< pos-current pos))))
+            (setq pos-best pos)
+            (setq pi-best pi)))))
+
+    pi-best))
+
+(defun bookmark-in-project--step-the-buffer (pi-list direction filepath-current)
+  "Step into a buffer in PI-LIST along DIRECTION relative to FILEPATH-CURRENT.
+Returning the next bookmark or nil."
+  (let*
+    (
+      (pi-best nil)
+      (filepath-pi-list-pairs
+        (bookmark-in-project--pretty-items-by-file-vector pi-list (list filepath-current)))
+      (files-len (length filepath-pi-list-pairs))
+      ;; No need to check if `i-best' is non-nil as the file was explicitly added.
+      (i-best
+        (bookmark-in-project--nth-vec filepath-current filepath-pi-list-pairs
+          :test (lambda (self other) (string-equal self (car other))))))
+
+    ;; It's possible there is only one buffer, in that case do nothing.
+    ;; We could check the file exists.
+    (let ((i (mod (+ i-best direction) files-len)))
+      (pcase-let ((`(,filepath-next . ,pi-list-local) (aref filepath-pi-list-pairs i)))
+        (setq pi-best
+          (bookmark-in-project--step-in-buffer
+            pi-list-local direction filepath-next
+            (cond
+              ((< direction 0)
+                most-positive-fixnum)
+              (t
+                0))))))
+
+    pi-best))
+
+(defun bookmark-in-project--step-any-impl (pi-list direction filepath-current pos-current)
+  "Step along DIRECTION in PI-LIST relative to FILEPATH-CURRENT & POS-CURRENT."
+  (or
+    (bookmark-in-project--step-in-buffer pi-list direction filepath-current pos-current)
+    (bookmark-in-project--step-the-buffer pi-list direction filepath-current)
+    ;; When there is only a single buffer, wrap back around to the start.
+    (bookmark-in-project--step-in-buffer
+      pi-list direction filepath-current
+      (cond
+        ((< direction 0)
+          most-positive-fixnum)
+        (t
+          0)))))
+
+(defun bookmark-in-project--step-any (pi-list direction)
+  "Step along DIRECTION in PI-LIST relative to the current buffer & position."
+  (let
+    (
+      (keep-searching t)
+      (skip 0)
+      (pi-best nil)
+      (pi-list-length (length pi-list))
+      (filepath-current (bookmark-in-project--canonicalize-path (buffer-file-name)))
+      (pos-current
+        (cond
+          ((< direction 0)
+            (line-beginning-position))
+          (t
+            (line-end-position)))))
+
+    (while (and keep-searching (< skip pi-list-length))
+      (let
+        ((pi (bookmark-in-project--step-any-impl pi-list direction filepath-current pos-current)))
+        (cond
+          (pi
+            ;; In practice this will be true, call this more for correctness.
+            ;; To ensure values are properly set.
+            (bookmark-in-project--pretty-item-from-bookmark pi)
+
+            (pcase-let ((`(,item . (,pos . ,invalid)) pi))
+              (cond
+                (invalid
+                  (setq skip (1+ skip))
+                  (setq filepath-current (bookmark-in-project--item-get-filename item))
+                  (setq pos-current pos))
+                (t
+                  (setq pi-best pi)
+                  (setq keep-searching nil)))))
+          (t
+            (setq keep-searching nil)))))
+    (cons pi-best skip)))
+
+(defun bookmark-in-project--calc-global-index (pi-list pi)
+  "Calculate the global index for PI in PI-LIST."
+  (let*
+    (
+      (filepath-item (bookmark-in-project--item-get-filename (car pi)))
+      (filepath-pi-list-pairs (bookmark-in-project--pretty-items-by-file-vector pi-list))
+      (filepath-bm-list-pairs-len (length filepath-pi-list-pairs))
+      (index 0)
+      (i 0))
+
+    (while (< i filepath-bm-list-pairs-len)
+      (pcase-let ((`(,filepath-iter . ,pi-list-iter) (aref filepath-pi-list-pairs i)))
+        (cond
+          ;; The file pas been met, find the index within the list.
+          ((string-equal filepath-iter filepath-item)
+            (setq pi-list-iter (bookmark-in-project--pi-list-sorted-by-pos pi-list-iter))
+            ;; The result is never nil, no need to account for this case.
+            (setq index (+ index (bookmark-in-project--nth-elt pi pi-list-iter)))
+            ;; Break.
+            (setq i filepath-bm-list-pairs-len))
+
+          ;; Not a matching file, step onto the next (accumulating index).
+          (t
+            (setq index (+ index (length pi-list-iter))))))
+      ;; Step over the file-paths.
+      (setq i (1+ i)))
+    index))
 
 (defun bookmark-in-project--default-name-at-point ()
   "Return the default name to use (based on surrounding context)."
@@ -530,27 +785,44 @@ otherwise it will switch the buffer."
       (t
         bookmark-current-bookmark))))
 
+
+;; ---------------------------------------------------------------------------
+;; Bookmark Jumping Logic
+
 (defun bookmark-in-project--jump-direction-impl (proj-dir bm-list direction)
   "Step bookmark in DIRECTION direction in PROJ-DIR & BM-LIST."
-  (setq bm-list (sort bm-list #'bookmark-in-project--compare))
-  (let*
-    (
-      (i-next (bookmark-in-project--step-index bm-list direction))
-      (item-next (nth i-next bm-list))
-      (name (car item-next)))
 
+  (let ((pi-list (bookmark-in-project--pretty-items-from-list bm-list)))
+    ;; Track the number of failed attempts (report this when verbose).
+    (pcase-let ((`(,pi . ,skip) (bookmark-in-project--step-any pi-list direction)))
+      (cond
+        (pi
+          (pcase-let ((`(,item . ,_pi-pos-cell) pi))
+            ;; Call jump (non-interactively).
+            (bookmark-jump item)
 
-    ;; Call jump (non-interactively).
-    (bookmark-jump item-next)
-
-    (when bookmark-in-project-verbose-cycle
-      (bookmark-in-project--message "(%s) %d of %d: %s"
-        (cond
-          ((< direction 0)
-            "prev")
-          (t
-            "next"))
-        (1+ i-next) (length bm-list) (bookmark-in-project--name-abbrev-and-fontify proj-dir name)))))
+            (when bookmark-in-project-verbose-cycle
+              (let ((name (car item)))
+                (bookmark-in-project--message "(%s) %d of %d: %s%s"
+                  (cond
+                    ((< direction 0)
+                      "prev")
+                    (t
+                      "next"))
+                  (1+ (bookmark-in-project--calc-global-index pi-list pi))
+                  (length bm-list)
+                  (bookmark-in-project--name-abbrev-and-fontify proj-dir name)
+                  (cond
+                    ((zerop skip)
+                      "")
+                    (t
+                      (format " (%d skipped)" skip))))))))
+        (t
+          ;; Note that his is  unlikely.
+          (when bookmark-in-project-verbose-cycle
+            (bookmark-in-project--message "unable to cycle %d bookmarks in %S!"
+              (length bm-list)
+              proj-dir)))))))
 
 (defun bookmark-in-project--jump-direction (direction)
   "Jump between bookmarks in DIRECTION (+1/-1)."